+++ Assets/Scenes/GameScene.unity
@@ -998,7 +998,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1121498765}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9f4147b7e6701479eb94f0f43f3ab9d6, type: 3}
+  m_Script: {fileID: 11500000, guid: 68df3e8bb9a2c43d484dc534840a32d8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _dropdown: {fileID: 710254189}
@@ -1169,7 +1169,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1829522086}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9da479f47fcc1284aa284be29167296d, type: 3}
+  m_Script: {fileID: 11500000, guid: f29290a3f74434701a501c7c7efbd443, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _triggerSound: {fileID: 8300000, guid: 485d769097c4ec44986ebf117315c745, type: 3}
