+++ Assets/Scripts/Mission.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b497ccfeda0f41ecb7c21a1c8261120
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
