+++ Assets/Scripts/Quests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f14497658048469385fa34a930a2114
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
