+++ Assets/Scripts/Quests/AudioTrigger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9f4147b7e6701479eb94f0f43f3ab9d6
+guid: f29290a3f74434701a501c7c7efbd443
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
