+++ Assets/Scripts/Quests/Mission.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c3b1e9ab70f82489384c80dd5e7c753d
+guid: 645d05b98ffc14efc9ffa93e98f0b855
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
