+++ Assets/Scripts/Quests/SlopeQuestTrigger.cs
@@ -1,9 +1,10 @@
 using System.Collections;
 using System.Collections.Generic;
-using Scripts.Game;
 using UnityEngine;
+using Scripts.Game;
+using Scripts.MissonLogMenu;
 
-namespace Scripts.MissonLogMenu
+namespace Scripts.Quests
 {
     public class SlopeQuestTrigger : MonoBehaviour
     {
