+++ Assets/Scripts/Quests/SlopeQuestTrigger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9da479f47fcc1284aa284be29167296d
+guid: 68df3e8bb9a2c43d484dc534840a32d8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
